requirements.txt
asn1crypto==1.5.1
attrs==21.4.0
beautifulsoup4==4.11.1
cffi==1.15.0
coincurve==17.0.0
enum34==1.1.10
gevent==21.12.0
gipc==1.4.0
gmpy2==2.1.2
greenlet==1.1.2
grpcio==1.46.1
grpcio-tools==1.46.1
importlib-metadata==4.11.3
iniconfig==1.1.1
lxml==4.8.0
packaging==21.3
pluggy==1.0.0
protobuf==3.20.1
py==1.11.0
pycparser==2.21
pycrypto==2.6.1
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
six==1.16.0
soupsieve==2.3.2.post1
tomli==2.0.1
typing_extensions==4.2.0
xlrd==2.0.1
zfec==1.5.7.2
zipp==3.8.0
zope.event==4.5.0
zope.interface==5.4.0
